# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/providers": {
      "get": {
        "operationId": "authProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Providers"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Which sign-in providers this server offers.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sign-in-code": {
      "post": {
        "operationId": "authRequestSignInCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingSignIn"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Send a sign-in code, once the Turnstile token passes. Same answer whether or not the address has an account.\nWithin the resend cooldown nothing is sent, and the answer says so (sent: false).",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sign-in-code/resend": {
      "post": {
        "operationId": "authResendSignInCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendSignInCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingSignIn"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Send another code to the address a pending token names.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sign-in": {
      "post": {
        "operationId": "authSignIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Sign in with a code; sets the session cookie. A first sign-in creates the account.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/social/{provider}/sign-in": {
      "post": {
        "operationId": "authSocialSignIn",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "discord",
                "github"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialSignInRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderRedirect"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Start signing in with Discord or GitHub: the address to send the browser to.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/social/{provider}/link": {
      "post": {
        "operationId": "authSocialLink",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "discord",
                "github"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderRedirect"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Start connecting Discord or GitHub to the signed-in account.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/social/{provider}": {
      "delete": {
        "operationId": "authSocialUnlink",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "discord",
                "github"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Disconnect Discord or GitHub. Needs a recent sign-in.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/session": {
      "get": {
        "operationId": "authSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "The signed-in account; 401 when there is none.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sign-out": {
      "post": {
        "operationId": "authSignOut",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Sign out; clears the session cookie.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sign-in-methods": {
      "get": {
        "operationId": "authSignInMethods",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInMethods"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Discord and GitHub: offered here, and attached to the signed-in account.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/email-change": {
      "post": {
        "operationId": "authRequestEmailChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailChangeStarted"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Send a code to a new address. Needs a recent sign-in.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/email-change/confirm": {
      "post": {
        "operationId": "authConfirmEmailChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailChangeConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Switch to the new address with the code sent to it.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/account": {
      "delete": {
        "operationId": "authDeleteAccount",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Delete the signed-in account and everything it owns. Needs a recent sign-in; clears the session cookie.",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "vrc.page API",
    "description": "The backend for vrc.page. Errors use RFC 9457 Problem Details (application/problem+json), and every response carries an X-Request-Id header.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "Providers": {
        "type": "object",
        "properties": {
          "discord": {
            "type": "boolean"
          },
          "github": {
            "type": "boolean"
          }
        },
        "required": [
          "discord",
          "github"
        ]
      },
      "SignInCodeRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The address to send the code to."
          },
          "botCheckToken": {
            "type": "string",
            "description": "The Turnstile token from the sign-in form's widget."
          }
        },
        "required": [
          "email",
          "botCheckToken"
        ]
      },
      "PendingSignIn": {
        "type": "object",
        "properties": {
          "pendingToken": {
            "type": "string",
            "description": "The address, dated when its code went out and signed:\n`<sent at ms>.<signature>.<email>`. Keep it (the website uses an httpOnly\ncookie) and hand it back to resend without another bot check. Valid as\nlong as the code."
          },
          "sent": {
            "type": "boolean",
            "description": "False inside the resend cooldown: the code sent moments ago still stands."
          },
          "resendIn": {
            "type": "number",
            "description": "Seconds before another code can be sent."
          }
        },
        "required": [
          "pendingToken",
          "sent",
          "resendIn"
        ]
      },
      "ResendSignInCodeRequest": {
        "type": "object",
        "properties": {
          "pendingToken": {
            "type": "string"
          }
        },
        "required": [
          "pendingToken"
        ]
      },
      "SignInRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "The digits from the email; anything else in it is ignored."
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "SocialSignInRequest": {
        "type": "object",
        "properties": {
          "callbackURL": {
            "type": "string",
            "description": "A path on the website, such as /dashboard."
          },
          "errorCallbackURL": {
            "type": "string",
            "description": "A path on the website; the provider's `error` is added to its query."
          }
        },
        "required": [
          "callbackURL",
          "errorCallbackURL"
        ]
      },
      "ProviderRedirect": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
      "SocialLinkRequest": {
        "type": "object",
        "properties": {
          "callbackURL": {
            "type": "string",
            "description": "A path on the website; on failure it gets an `error` query parameter."
          }
        },
        "required": [
          "callbackURL"
        ]
      },
      "SessionInfo": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "description": "The account's own address, for its settings. Never shown anywhere public."
          }
        },
        "required": [
          "accountId",
          "email"
        ]
      },
      "SignInMethod": {
        "type": "object",
        "properties": {
          "configured": {
            "type": "boolean",
            "description": "Offered on this server."
          },
          "connected": {
            "type": "boolean",
            "description": "Attached to the signed-in account."
          }
        },
        "required": [
          "configured",
          "connected"
        ]
      },
      "SignInMethods": {
        "type": "object",
        "properties": {
          "discord": {
            "$ref": "#/components/schemas/SignInMethod"
          },
          "github": {
            "$ref": "#/components/schemas/SignInMethod"
          }
        },
        "required": [
          "discord",
          "github"
        ]
      },
      "EmailChangeRequest": {
        "type": "object",
        "properties": {
          "newEmail": {
            "type": "string"
          }
        },
        "required": [
          "newEmail"
        ]
      },
      "EmailChangeStarted": {
        "type": "object",
        "properties": {
          "resendIn": {
            "type": "number",
            "description": "Seconds before another code can be sent to this address."
          }
        },
        "required": [
          "resendIn"
        ]
      },
      "EmailChangeConfirmRequest": {
        "type": "object",
        "properties": {
          "newEmail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "newEmail",
          "code"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "The HTTP status code.",
            "example": 404
          },
          "retryAfter": {
            "type": "integer",
            "description": "Seconds to wait before trying again, for a cooldown."
          },
          "at": {
            "type": "integer",
            "description": "Which item of a submitted list the problem is about, counting from 0."
          },
          "type": {
            "type": "string",
            "description": "A URI naming the kind of problem: \"about:blank\" when the status says it\nall, otherwise https://vrc.page/problems/<code>, with a code from\nsrc/common/problem.ts."
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the kind of problem."
          },
          "detail": {
            "type": "string",
            "description": "What went wrong this time, in words a person can act on."
          },
          "instance": {
            "type": "string",
            "description": "The path of the request that failed."
          },
          "requestId": {
            "type": "string",
            "description": "Same as the X-Request-Id response header and the request's audit log entries."
          }
        },
        "required": [
          "status",
          "type",
          "title",
          "requestId"
        ]
      }
    }
  }
}